add 1 balancer to simulate consistent external IP
the balancer has kube-proxy and joins the overlay network, so it's able to
accept requests to a service and forward it to correct pods, but otherwise will
not be hosting pods / containers.

this helps simulate the usecase where we want to have only 1 consistent external
IP to a service locally instead of having to use 1 minion node as the balancer.

Signed-off-by: Daniel, Dao Quang Minh <dqminh89@gmail.com>